requirements.txt
# Core dependencies
chromadb>=0.4.18
pypdf>=3.16.0
sentence-transformers>=2.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
requests>=2.31.0

# Data processing
PyYAML>=6.0.1
numpy>=1.24.0
pandas>=2.1.1
langchain>=0.0.335

# Utilities
tqdm>=4.66.1
python-dotenv>=1.0.0
jsonlines>=3.1.0

# Testing
pytest>=7.4.2
